requirements.txt
fastapi
uvicorn
python-dotenv
solana
requests
pydantic
